create separate tests
