+++ polls/tests/test_poll_dates.py
@@ -1,0 +1,74 @@
+import datetime
+
+from django.test import TestCase
+from django.utils import timezone
+
+from polls.models import Question
+
+
+def create_question(question_text, days):
+    """
+    Create a question with the given `question_text` and published the
+    given number of `days` offset to now (negative for questions published
+    in the past, positive for questions that have yet to be published).
+    """
+    time = timezone.now() + datetime.timedelta(days=days)
+    return Question.objects.create(question_text=question_text, pub_date=time)
+
+
+class QuestionModelDateTests(TestCase):
+
+    def test_was_published_recently_with_future_question(self):
+        """
+        was_published_recently() returns False for questions whose pub_date
+        is in the future.
+        """
+        time = timezone.now() + datetime.timedelta(days=30)
+        future_question = Question(pub_date=time)
+        self.assertFalse(future_question.was_published_recently())
+
+    def test_was_published_recently_with_old_question(self):
+        """
+        was_published_recently() returns False for questions whose pub_date
+        is older than 1 day.
+        """
+        time = timezone.now() - datetime.timedelta(days=1, seconds=1)
+        old_question = Question(pub_date=time)
+        self.assertFalse(old_question.was_published_recently())
+
+    def test_was_published_recently_with_recent_question(self):
+        """
+        was_published_recently() returns True for questions whose pub_date
+        is within the last day.
+        """
+        time = timezone.now() - datetime.timedelta(hours=23,
+                                                   minutes=59, seconds=59)
+        recent_question = Question(pub_date=time)
+        self.assertTrue(recent_question.was_published_recently())
+
+    def test_is_published_with_future_pub_date(self):
+        """
+        is_published() returns False for questions whose pub_date
+        is in the future.
+        """
+        time = timezone.now() + datetime.timedelta(days=30)
+        future_question = Question(pub_date=time)
+        self.assertFalse(future_question.is_published())
+
+    def test_is_published_with_default_pub_date(self):
+        """
+        is_published() returns True for questions whose pub_date
+        is in the default pub date (now).
+        """
+        time = timezone.now()
+        now_question = Question(pub_date=time)
+        self.assertTrue(now_question.is_published())
+
+    def test_is_published_with_past_pub_date(self):
+        """
+        is_published() returns True for questions whose pub_date
+        is in the past.
+        """
+        time = timezone.now() - datetime.timedelta(days=30)
+        past_question = Question(pub_date=time)
+        self.assertTrue(past_question.is_published())
